- Finalizada a implementação da home.

+++ BankApp/app/src/main/java/br/com/rms/bankapp/data/local/database/entity/Statement.kt
@@ -10,5 +10,5 @@ data class Statement(
     val title: String?,
     val desc: String?,
     val date: String?,
-    val value: Int?
+    val value: Double?
 )
